api/refresh/requirements.txt
requests==2.31.0
google-cloud-firestore>=2.14.0
